% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSignificanceLevel.R
\name{GetConfidenceLevel}
\alias{GetConfidenceLevel}
\title{Used internally by all functions using alpha. Generates a string in the form of 100(1-alpha) percent}
\usage{
GetConfidenceLevel(alpha)
}
\description{
Used internally by all functions using alpha. Generates a string in the form of 100(1-alpha) percent
}
\keyword{internal}
