% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalOutput.R
\name{IntervalOutput}
\alias{IntervalOutput}
\title{Used internally by all functions producing intervals of any sort (prediction, confidence)}
\usage{
IntervalOutput(xbar, marginOfError, confidence,
  confidenceOrPrediction = "Confidence", one.sided)
}
\description{
Used internally by all functions producing intervals of any sort (prediction, confidence)
}
\keyword{internal}
