% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.2Pop.Hypothesis.R
\name{Mean.2Pop.Hypothesis}
\alias{Mean.2Pop.Hypothesis}
\title{Two Population Means - Hypothesis Test}
\usage{
Mean.2Pop.Hypothesis(xbar1, xbar2, n1, n2, var1, var2,
  differenceInPopulationMeans = 0, popVarKnown = F,
  alternative = "two.sided", equal = "F", alpha = 0.05)
}
\arguments{
\item{xbar1}{the sample mean of population 1}

\item{xbar2}{the sample mean of population 2}

\item{n1}{the size of the sample from population 1}

\item{n2}{the size of the sample from population 2}

\item{var1}{the sample or population variance of population 1 (the latter only if it is known).}

\item{var2}{the sample or population variance of population 2 (the latter only if it is known).}

\item{differenceInPopulationMeans}{the hypothesized difference between the population means. Set to 0 by default.}

\item{popVarKnown}{boolean denoting whether the population variance is known. False by default.}

\item{alternative}{string denoting the form of the alternative hypothesis Acceptable values are "smaller" (<), "greater" (>), and "one.sided" (=/=).}

\item{equal}{boolean denoting whether the population variances are equal. False by default. Used only in the case when the population variances are unknown.}

\item{alpha}{the significance level. Set to 0.05 by default.}
}
\value{
The test statistic, p-value, and conclusion of the hypothesis.
}
\description{
Performs a hypothesis test with 100(1-alpha) percent confidence for the difference in the means of two populations
}
