% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.1Regressor.CI.R
\name{Slope.1Regressor.CI}
\alias{Slope.1Regressor.CI}
\title{Simple Linear Regression - Slope Confidence Interval}
\usage{
Slope.1Regressor.CI(form, data, n, alpha = 0.05, one.sided = F)
}
\arguments{
\item{form}{expression representing the relationship between the variables contained in data (e.g., y ~ x). Variable names in this expression must match the column names of data.}

\item{data}{data frame object containing the data. Columns must represent the predictor and response variables, in either order.}

\item{n}{the number of data points collected.}

\item{alpha}{the significance level. Set to 0.05 by default.}

\item{one.sided}{boolean denoting whether the function should return one-sided confidence intervals. False by default.}
}
\value{
The confidence interval(s): one-sided lower and upper bounds (in a list) or just the two-sided confidence interval (as a simple vector).
}
\description{
Constructs a 100(1-alpha) percent confidence interval for the slope of a simple linear regression model.
}
