% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.1Regressor.Hypothesis.R
\name{Slope.1Regressor.Hypothesis}
\alias{Slope.1Regressor.Hypothesis}
\title{Simple Linear Regression - Hypothesis Test}
\usage{
Slope.1Regressor.Hypothesis(form, data, beta_hypothesized, n,
  alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{form}{expression representing the relationship between the variables contained in data (e.g., y ~ x). Variable names in this expression must match the column names of data.}

\item{data}{data frame object containing the data. Columns must represent the predictor and response variables, in either order.}

\item{beta_hypothesized}{the hypothesized value of the slope of the true linear model for the data, as specified by the null hypothesis.}

\item{n}{the number of data points collected.}

\item{alternative}{string denoting the form of the alternative hypothesis Acceptable values are "smaller" (<), "greater" (>), and "one.sided" (=/=).}

\item{alpha}{the significance level. Set to 0.05 by default.}
}
\value{
The test statistic, p-value, and conclusion of the hypothesis.
}
\description{
Performs a hypothesis test with 100(1-alpha) percent confidence for the slope of a simple linear regression model.
}
