% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.1Pop.CI.R
\name{Variance.1Pop.CI}
\alias{Variance.1Pop.CI}
\title{One Population Variance - Confidence Interval}
\usage{
Variance.1Pop.CI(sampleVar, n, alpha = 0.05, one.sided = F)
}
\arguments{
\item{sampleVar}{the variance of the sample}

\item{n}{the size of the sample}

\item{alpha}{the significance level. Set to 0.05 by default.}

\item{one.sided}{boolean denoting whether the function should return one-sided confidence intervals. False by default.}
}
\value{
The confidence interval(s): one-sided lower and upper bounds (in a list) or just the two-sided confidence interval (as a simple vector).
}
\description{
Constructs a 100(1-alpha) percent confidence interval for a single population variance
}
