% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.1Pop.CI.R
\name{Mean.1Pop.CI}
\alias{Mean.1Pop.CI}
\title{Construct a 100(1-alpha) percent confidence interval for a single population mean}
\usage{
Mean.1Pop.CI(xbar, variance, n, popVarKnown = F, alpha = 0.05,
  one.sided = F)
}
\arguments{
\item{xbar}{The mean of the sample}

\item{variance}{Either the sample or population variance (the latter only if it is known)}

\item{n}{The size of the sample}

\item{popVarKnown}{Boolean denoting whether the population variance is known. False by default.}

\item{alpha}{The significance level. Set to 0.05 by default.}

\item{one.sided}{Boolean denoting whether the function should return one-sided confidence intervals. False by default.}
}
\value{
The confidence interval(s): one-sided lower and upper bounds or just the two-sided confidence interval.
}
\description{
Construct a 100(1-alpha) percent confidence interval for a single population mean
}
