% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.1Pop.Hypothesis.R
\name{Mean.1Pop.Hypothesis}
\alias{Mean.1Pop.Hypothesis}
\title{Conduct a hypothesis test with 100(1-alpha) percent confidence for the mean of a single population}
\usage{
Mean.1Pop.Hypothesis(xbar, mu, variance, n, popVarKnown = T,
  alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{xbar}{The mean of the sample.}

\item{mu}{The hypothesized value of the population mean.}

\item{variance}{Either the sample or population variance (the latter only if it is known).}

\item{n}{The size of the sample.}

\item{popVarKnown}{Boolean denoting whether the population variance is known. False by default.}

\item{alternative}{A string denoting the form of the alternative hypothesis Acceptable values are "smaller" (<), "greater" (>), and "one.sided" (=/=).}

\item{alpha}{The significance level. Set to 0.05 by default.}
}
\value{
The pvalue produced by the hypothesis test.
}
\description{
Conduct a hypothesis test with 100(1-alpha) percent confidence for the mean of a single population
}
