% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.1Pop.PredictionInterval.R
\name{Mean.1Pop.PredictionInterval}
\alias{Mean.1Pop.PredictionInterval}
\title{Construct a 100(1-alpha) percent prediction interval for the future observation of a single population's mean}
\usage{
Mean.1Pop.PredictionInterval(xbar, variance, n, alpha = 0.05,
  popVarKnown = F, one.sided = F)
}
\arguments{
\item{xbar}{The mean of the sample.}

\item{variance}{Either the sample or population variance (the latter only if it is known).}

\item{n}{The size of the sample.}

\item{alpha}{The significance level. Set to 0.05 by default.}

\item{popVarKnown}{Boolean denoting whether the population variance is known. False by default.}

\item{one.sided}{Boolean denoting whether the function should return one-sided confidence intervals. False by default.}
}
\value{
The confidence interval(s): one-sided lower and upper bounds (in a list) or just the two-sided confidence interval (as a simple vector).
}
\description{
Construct a 100(1-alpha) percent prediction interval for the future observation of a single population's mean
}
